Added prototypical switch.
In pure Grace.
Have not yet tested for performance.
Have not yet written tests of correctness.
(It seems to basically work from a couple of simple examples,
but we still need to actually write tests for it).
The API for this is to set up a switch with a variadic param. list,
which constructs an object containing a sorted array.
This uses Unix qsort in the C implementation,
so it's worst case O(n^2)... to set up the switch.
Since setting up the switch is done at runtime,
and requires at least O(n) and worst case O(n^2) time,
it's only worth using if the same switch is accessed many times.
(probably).