Either way, an error (lexer).
I think the return value here *should* be false...
since it was initialized to false and not set...
but there was some (annoying, implicit)
statefulness in the ifelses.
Or not state, but the threat of it.
I may have missed something.
Anyway, I think false is right -
so this commit is maybe more wrong -
but interesting to check out and compare.
Methinks the actual problem is somewhere else in lexer,
though, or maybe in switch -
we should unit test switch (since that's easy)
and hypothetically I guess lexer,
but you heard Dr. Black note that
it's basically acceptance tested because
unit testing has just not been feasible yet.
So...