Refactored (but unfortunately slower) lexer.
The token-processing pass now takes 1.5 times as long.

Nonetheless this serves as an example of the syntax and algorithm for
using the switch.

I haven't yet profiled this to figure out why it's slower,
but there could be any number of reasons - method overhead,
the runtime array sorting, etc.

How can we make this switch faster?
Perhaps by modifying AST, genc (and eventually genjs)?