Added switch()case{} statement
Backed out Forrest's lexer changes for now (collisions with
my implementation).

Implemented the ast, parser, and identifierresolution elements of switch()case{}.

Syntax matches match()case{}, and the implementation is identical (outside of ast,
and parser, etc.).  The implementation of the actually switch routine still needs
to be put together, but gets parsed successfully so I think it's a step in the
right direction

We can reconcile Forrest and my implementations; Forrest's had the correct
algorithm (binary search), while this is parsed into the ast and can be used
by any grace program without including any modules.