Added switchcase tests.
The current match-like implementation succeeds for expressions,
explicit exceptions used for any case, etc.

However it currently fails one test:
when not even the _ case is defined, switch just returns an empty block
which is undesireable - I think it should be an exception,
because, especially when switch is used for expressions,
not meeting any of the cases is a potentially serious issue.
This exception can always be remedied with a _ default case.