Shares code between two versions of fresh methods.
Previously, fresh methods were compiled twice, once for use in
requests and again for use in inherits and uses.  This doubled the
size of the compiled code for every class.  With this change, the
version of the method for use in requests simply creates a fresh
graceObject and then uses the $build(3) and $init(1) methods to
populate it.