Adds an asString method to the test module.
The default method is being overridden by the inherited
asString method in prelude.method.  This is the correct
behaviour.  So to check that self is correct, we need to
define a local asString method. This  makes the js and c
tests identical, so it makes sense to share the same files.